Remove float32 casting from std_gaussian_centres

+++ craystack/codecs.py
@@ -508,8 +508,7 @@ def std_gaussian_centres(precision):
     if precision in std_gaussian_centres_cache:
         return std_gaussian_centres_cache[precision]
     else:
-        centres = np.float32(
-            norm.ppf((np.arange(1 << precision) + 0.5) / (1 << precision)))
+        centres = norm.ppf((np.arange(1 << precision) + 0.5) / (1 << precision))
         std_gaussian_centres_cache[precision] = centres
         return centres
